Fix unstopable thread issue. Fix direction issue.
