+++ world-simulator/src/main/java/com/example/worldsimulator/simulator/TargetSimulatorRunner.java
@@ -1,7 +1,11 @@
 package com.example.worldsimulator.simulator;
 
 import org.springframework.boot.CommandLineRunner;
+import org.springframework.context.annotation.Bean;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.stereotype.Component;
+
+import java.util.concurrent.Executor;
 
 @Component
 public class TargetSimulatorRunner  implements CommandLineRunner {
@@ -15,4 +19,18 @@ public class TargetSimulatorRunner  implements CommandLineRunner {
     public void run(String... args) throws Exception {
         simulator.start();
     }
+
+
+    @Bean
+    public Executor taskExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(1);
+        executor.setMaxPoolSize(1);
+        executor.setDaemon(true);
+        executor.setThreadNamePrefix("WorldSimulator-");
+
+        executor.initialize();
+        return executor;
+    }
+
 }
