+++ world-simulator/src/main/resources/application.properties
@@ -1,5 +1,8 @@
 spring.application.name=world-simulator
 server.port=8083
+server.shutdown=immediate
+spring.lifecycle.timeout-per-shutdown-phase=50ms
+
 spring.kafka.bootstrap-servers= localhost:9092
 spring.kafka.producer.key-serializer= org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer= org.springframework.kafka.support.serializer.JsonSerializer
@@ -7,5 +10,5 @@ spring.kafka.topic.name=TargetPointPosition
 target.heading=360
 target.initial.position.X = -1
 target.initial.position.Y = 5
-target.speed=0.1
+target.speed=0.5
 target.name=redApple
